requirements.txt
pygame
typing_extensions
pyglm==2.7.3
loguru